Watcher API functions (#10)
* Stubbed out Watcher Commands

All watcher api endpoints should be stubbed out as functions under
escmd. Only escmd.WatcherList is fully implemented so far (since it
requires no arguments and was the most fun to add).
In addition to this, a debug snippet was added to escfg. It needs to be
wrapped around a feature flag and should be delivered once
implementation details can be determined.
Finally, TODO.md was updated to track progress of the Watcher commands.

* Added watcher functions

* Updated command list. Added comments