Reduced errors in output format feature
In the previous iteration of this feature, an if statement was used to
assign a value to the Format field of a struct. The if statement used
reflection to determine whether or not that field existed before
assignment. In cases where the field did not exist, the code did not
compile. Instead a solution was created where a pointer to the original
struct was passed to esutil.SetFormat, which uses reflection to
correctly handle cases where this field does not exist.

escmd/cat.go and escmd/get.go are now able to use this feature flag.
Future work will improve esutil.ParseBytes to handle cases where the
Format field does not exist, yet the user wants a specific output
format. Is it worth asking Elastic to standardize on a Format field for
every api endpoint?