Integrated some watcher funcs w output feature
Some watcher functions have added support for the output format flag.
Unfortunately, the watcher functions were not fully implemented prior to
this feature, so there is not a consistent state. Still dropping a
commit so the progress can be tracked (and so other modules can be
worked on in parallel)