Feature/output format (#11)
* Added support for -o flag. Controls output format

This commit adds preliminary support for the -o flag to allow output of
yaml or json output formats. Functions seeking to enable this should use
RequestNew for now. Once existing code has been ported to consistently
use RequestNew, the Request function signature will be deprecated in
favor of RequestNew's signature.

The cat command has been ported to this new format

* Reduced errors in output format feature

In the previous iteration of this feature, an if statement was used to
assign a value to the Format field of a struct. The if statement used
reflection to determine whether or not that field existed before
assignment. In cases where the field did not exist, the code did not
compile. Instead a solution was created where a pointer to the original
struct was passed to esutil.SetFormat, which uses reflection to
correctly handle cases where this field does not exist.

escmd/cat.go and escmd/get.go are now able to use this feature flag.
Future work will improve esutil.ParseBytes to handle cases where the
Format field does not exist, yet the user wants a specific output
format. Is it worth asking Elastic to standardize on a Format field for
every api endpoint?

* Integrated some watcher funcs w output feature

Some watcher functions have added support for the output format flag.
Unfortunately, the watcher functions were not fully implemented prior to
this feature, so there is not a consistent state. Still dropping a
commit so the progress can be tracked (and so other modules can be
worked on in parallel)