Added command support for user/pass in cfg file
This commit focuses on delivering the ability to execute arbitrary
commands to obtain credentials used to authenticate to an elasticsearch
cluster. This has its risks, but allows users to add a password manager
as a layer to prevent credentials stored in a config file (requested by
@slmingol).

There is a lot of new code in this patch, so it will likely sit on a
feature branch for a few days until tests can be developed or users can
provide feedback.