(partial) added support for --help flag
The first funtions that are being ported from escli are help_cat and
help_indices. Functions like those don't really scale well, so a flag is
being added to _all_ other commands (including the generic request
command) to support helping in a scalable way. Now, there only needs to
be modification to the boilerplate to the set of implemented functions
rather than mocking adding dedicated commands for every thing that needs
to be explained. This is partial because the change is only implemented
in the cmd package (pkg/cmd). Integration with existing packages will
come a bit later